Update README.md
This repo contains two R scripts - one to create the data visualizations to be used in the example Shiny App (DPdataprep.r) & one to implement the App (DPapp.r). Any required packages needed to load, process, clean, and visualized the data should be automatically downloaded as a process of executing the code. This app is intended to provide some example code to the public space on how to: access different sources of publicly available data without first downloading and storing files locally, clean and plot data in a tidy environment, and generate a fairly simple Shiny App.